requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
schedule>=1.2.0
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-CORS>=4.0.0
waitress>=2.1.0

# 可选依赖（用于Windows服务）
pywin32>=306; sys_platform == "win32"

# 开发依赖（可选）
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
